Reorganize Vassiliev, Update Vas_Defs_JW
Still trying to utilize the Vas_Defs_JW code in Vassiliev.py, but we have to get the code working first. Additionally, there is trouble with the original Vassiliev.py code where using the vas_open method on a closed knot gives the wrong answer even when the first point is appended onto the knot.