Update ScanAnalysis, VasFigures
Add plot_vas_change to ScanAnalysis.py, whose function is to graph the second Vassiliev measures of each section of each protein against the starting point of said section. This allows us to see how the V2 evolves along the chain of atoms, at least when considering the the local V2 of sections with length 200, 400, and 600.
Actual scan data is still incomplete, but the code to analyze the data is functional.
Additionally, this commit adds the last of the proteins' V2 measures at 100 projections for the purpose of thoroughness.