SpeedTest, Upload V2 Evolutions
Uploaded png files to Vas-Data/Change, ran speed test to conclude that using 4 processors as the parameter for multiprocessing.Pool is somehow faster than using more if there are more available.
Also obtained thorough scan results, which just include 6zge at 200 scanlength with an interval of 1. It is missing the last interval, but that is easy to fix.