ReProcess Change and Scans,
Reprocess Vas-Data/Change and Vas-Data/Scans png files based on new scans from Merge files. All that is missing is 6zgg, 7kdk, and the 600-projection scans for 6acd and 7m8k.
I have also deleted SpeedTest because I have concluded that on the cluster there is not a significant positive impact on performance when changing the number of processors parameter for multiprocessing.Pool closer to the total available (28) instead of simply 4.
Now I must analyze the new plots and see if there are interesting inferences that can be made from them based on the biological contexts surrounding the proteins respective to each of them.