Reorganize Files, Matrix Fix Attempt
Reorganized the Graph folder names and location of scan text files into a new folder. Also created three new text files for use in new plot_all_change method in ScanAnalysis.py.
Matrix script fixes, mainly deleting the 'if not i' statement on previous line 231, which was supposed to create a new row when necessary, seem to be working on my laptop but unsuccessful in stopping the odd behaviour on the cluster. It still will not print out any text output.