Add Cluster output, Update MatrixVis
Moved Scan-Graphs so I could redo them using the new amino_acid indices, but that code (in ScanAnalysis) still needs work.
Also figured out that the reason my matrix csv output was so weird was just because it was appending the whole dataframe to the csv after every i iteration (as opposed to just appending the next line). So that's chill, and it means my code basically works.
Also tried to dig through the confusing matrix visualization code (in the file now called MatrixVis.py), which was frustrating but not impossible once I figured out that only a small snippet of the code was actually for what I wanted which was just to visualize the matrix. Still more than a little buggy due to there being at least 3 different data structures being used for the same matrix, but it looks 1000x better.