Update Domain Structure Data, Attempt Graph Update
Attempted to update grouped graphs so that the section missing from 6zgh would be more obviously missing data, also rearranged the proteins in the Wild Proteins graphs (6zg*) so that the proteins are in the correct order.
Also added more complete domain structure info to both Vas-Figures and, in the form of the original source plot, to FinalPresentation.